wip: refactored useMerchantForm hook into a onboard directory and set of hooks for building form
